+++ agrest-cayenne/src/test/java/io/agrest/cayenne/GET/Exp_DisallowedIT.java
@@ -6,7 +6,6 @@ import io.agrest.cayenne.cayenne.main.E2;
 import io.agrest.cayenne.unit.main.MainDbTest;
 import io.agrest.cayenne.unit.main.MainModelTester;
 import io.agrest.jaxrs3.AgJaxrs;
-import io.agrest.meta.AgEntity;
 import io.bootique.junit5.BQTestTool;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
@@ -15,19 +14,15 @@ import jakarta.ws.rs.core.Configuration;
 import jakarta.ws.rs.core.Context;
 import org.junit.jupiter.api.Test;
 
-public class Exp_PropFilter_Overlay_RuntimeIT extends MainDbTest {
+public class Exp_DisallowedIT extends MainDbTest {
 
     @BQTestTool
     static final MainModelTester tester = tester(Resource.class)
             .entitiesAndDependencies(E2.class)
-            .agCustomizer(c -> c.entityOverlay(AgEntity.overlay(E2.class)
-                    .readablePropFilter(b -> b.property("name", false))
-                    .writablePropFilter(b -> b.property("name", false))))
             .build();
 
-    // TODO: is this correct - we allow filtering on a non-readable attribute?
     @Test
-    public void expOnHiddenProperty() {
+    public void dbExp() {
 
         tester.e2()
                 .insertColumns("id_", "name", "address")
@@ -36,10 +31,9 @@ public class Exp_PropFilter_Overlay_RuntimeIT extends MainDbTest {
                 .exec();
 
         tester.target("")
-                .queryParam("exp", "name = 'n1'")
+                .queryParam("exp", "db:id_ = 1")
                 .get()
-                .wasOk()
-                .bodyEquals(1, "{\"id\":1,\"address\":\"a1\"}");
+                .wasBadRequest();
     }
 
     @Path("")